Room Region analysis now has the same result across multiple runs.
Changed from the non-deterministic priority_queue dependency to
std::collections::BinaryHeap (which I originally didn't realize
existed). This also seems to give a slight performance improvement?
But it's on the order of maybe 20 microseconds, and I'd need to do
stats and run across every room to be sure.

Apparently cargo also makes you increment the minor version when you
remove a dependency.