Fix edge case for local maxima on border tile
Since the distance transform and maxima finding is independent of
the exit region setup, there was an edge case possible where the
local maxima for a non-exit region was in the location of what
would otherwise be a border tile between the exit region and the
non-exit region.

This update adds logic to exclude exit region tiles and positions
that will be their borders from the local maxima used for the
watershed region finding process.